base module update

+++ seims/src/seims_main/base/data/NotRegularMeasurement.cpp
@@ -73,7 +73,7 @@ NotRegularMeasurement::NotRegularMeasurement(mongoc_client_t *conn, string hydro
             }
 
             if (bson_iter_init(&iter, doc) && bson_iter_find(&iter, MONG_HYDRO_DATA_UTC)) {
-                dt = GetDatetimeFromBsonIterator(&iter) / 1000.f;
+                dt = (float) GetDatetimeFromBsonIterator(&iter) / 1000.f;
             } else {
                 throw ModelException("NotRegularMeasurement", "NotRegularMeasurement",
                                      "The UTCDateTime field does not exist in DataValues table.");
